% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_phoQP_hit_peptide.R
\name{imprints_phoQP_hit_peptide}
\alias{imprints_phoQP_hit_peptide}
\title{imprints_phoQP_hit_peptide}
\usage{
imprints_phoQP_hit_peptide(
  data,
  data_diff,
  ctrl,
  phospho = TRUE,
  FC_cutoff = 0.5,
  fixed_score_cutoff = TRUE,
  FDR = 0.01,
  curvature = 0.1,
  folder_name = ifelse(phospho, "PhosphoHits_analysis", "QPpeptideHits_analysis")
)
}
\arguments{
\item{data}{The input data set from which categorization is performed on and hitlist is produced from;
i.e. the peptide dataset after normalization}

\item{data_diff}{The peptide output from imprints_caldiff; can also be a path to the data.}

\item{ctrl}{The name of the control.}

\item{phospho}{Is your data of phosphoproteomics type. If not, should be quantiative proteomics.}

\item{FC_cutoff}{The fold change cutoff}

\item{fixed_score_cutoff}{Logical to tell if you want to use a fixed cutoff for the I-score.
If TRUE, the value IS_cutoff will directly be used as the cutoff and for all treatments. If FALSE,
the I-score cutoff will be calculated as the value selected for IS_cutoff plus the median of the
I-scores of the proteins which have a p-value lower than the median of all p-values for a given treatment.
Default is TRUE.}

\item{FDR}{The FDR used for the BH corrected combined p-value}

\item{curvature}{The curvature used for the curve on the volcano plot}

\item{folder_name}{The name of the folder in which you want to save the results.}
}
\value{
A dataframe which contains the hits.
}
\description{
Function to categorize peptide according to their Expression change
}
