% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_phoQP_join.R
\name{imprints_phoQP_join}
\alias{imprints_phoQP_join}
\title{imprints_phoQP_join}
\usage{
imprints_phoQP_join(
  phospho = NULL,
  QPpep = NULL,
  QP = NULL,
  method = c("full", "inner")
)
}
\arguments{
\item{phospho}{Phosphoproteomics data processed  after \code{imprints_phospho_rawread}}

\item{QPpep}{Quantitative proteomics data on the peptide level processed  after \code{imprints_QPpep_rawread}}

\item{QP}{The quantitative proteomics data on the protein level; it can also contains the CETSA data}

\item{method}{Either full or inner join}
}
\value{
The joined dataset
}
\description{
Function to join phosphoproteomics data, quantitative data on protein and peptide level
}
