% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_phoQP_ksea.R
\name{imprints_phoQP_ksea}
\alias{imprints_phoQP_ksea}
\title{imprints_phoQP_ksea}
\usage{
imprints_phoQP_ksea(data, save = FALSE)
}
\arguments{
\item{data}{The analysis tab output from the imprints_phoQP_hit_peptide 
(either a data frame or a path to the file)}

\item{save}{Logical to tell if you want to save the ksea file in the csv format for each treatment}
}
\value{
A data frame in the KSEA format
}
\description{
Function to put the analysis tab in a summary format suitable to use the \link[https://casecpb.shinyapps.io/ksea/]{KSEA app}
}
