% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_phoQP_pltformat.R
\name{imprints_phoQP_pltformat}
\alias{imprints_phoQP_pltformat}
\title{imprints_phoQP_pltformat}
\usage{
imprints_phoQP_pltformat(data, hits = NULL, reftreatment = NULL)
}
\arguments{
\item{data}{Processed data after \code{imprints_QPpep_rawread} or \code{imprints_phospho_rawread}}

\item{hits}{A vector of ids that you would like to keep}

\item{reftreatment}{A character corresponding to the reference treatment that you 
wish to remove (can be NULL)}
}
\value{
A filter data frame ready to use in \code{\link[IMPRINTS.CETSA.app:imprints_barplotting_app]{imprints_barplotting_app}}
}
\description{
Function to put the processed data after imprints_QPpep_rawread or imprints_phospho_rawread
in the right format to use \code{\link[IMPRINTS.CETSA.app:imprints_barplotting_app]{imprints_barplotting_app}}
}
