% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_QPpep_rawread.R
\name{imprints_QPpep_rawread}
\alias{imprints_QPpep_rawread}
\title{imprints_QPpep_rawread}
\usage{
imprints_QPpep_rawread(
  QP_file,
  treatment,
  prefixcontaminant = "",
  countthreshold = 1,
  dataset_name = "QPpeptide"
)
}
\arguments{
\item{QP_file}{The path to the PeptidesGroups file from PD.}

\item{treatment}{The treatment corresponding to each fo your TMT channel. It should be
the same length as your number of TMT channel and should follow the format 'Rep_treatment'
like 'B1_Vehicle' for example.}

\item{prefixcontaminant}{Character corresponding to the prefix used to identify contaminants.}

\item{countthreshold}{The minimal threshold number of associated abundance count of peptides, default is 1.}

\item{dataset_name}{The name of your dataset}
}
\value{
Your data filtered and in the right format to use the IMPRINTS.CETSA package.
}
\description{
Function to read the quantitative proteomics data on the peptide level from Proteome Discoverer
}
