% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_phoQP_hit_peptide.R
\name{imprints_phoQP_hit_peptide}
\alias{imprints_phoQP_hit_peptide}
\title{imprints_phoQP_hit_peptide}
\usage{
imprints_phoQP_hit_peptide(
  data,
  data_diff = NULL,
  ctrl,
  phospho = TRUE,
  FC_cutoff = 0.5,
  FDR = 0.01,
  curvature = 0.1,
  folder_name = ifelse(phospho, "PhosphoHits_analysis", "QPpeptideHits_analysis"),
  peptide_count_col = "peptides_counts_all",
  species = "Homo Sapiens"
)
}
\arguments{
\item{data}{The input data set from which categorization is performed on and hitlist is produced from;
i.e. the peptide dataset after normalization}

\item{data_diff}{The peptide output from imprints_caldiff; can be NULL and if so will compute it; can also be a path to the data.}

\item{ctrl}{The name of the control.}

\item{phospho}{Is your data of phosphoproteomics type. If not, should be quantiative proteomics.}

\item{FC_cutoff}{The fold change cutoff}

\item{FDR}{The FDR used for the BH corrected combined p-value}

\item{curvature}{The curvature used for the curve on the volcano plot}

\item{folder_name}{The name of the folder in which you want to save the results.}

\item{peptide_count_col}{The name of the column that contain the unique peptide count.
If it is sumUniPeps, don't bother with this parameter.}

\item{species}{The species on which you did the experiment (not necessary if already present in your datas).
Default is 'Homo Sapiens'.}
}
\value{
A dataframe which contains the hits.
}
\description{
Function to categorize peptide according to their Expression change
}
