% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_phoQP_hit_protein.R
\name{imprints_phoQP_hit_protein}
\alias{imprints_phoQP_hit_protein}
\title{imprints_phoQP_hit_protein}
\usage{
imprints_phoQP_hit_protein(
  data,
  data_diff = NULL,
  ctrl,
  QP_name = "36C",
  FC_cutoff = 0.2,
  FDR = 0.05,
  curvature = 0.1,
  folder_name = "QPHits_analysis",
  peptide_count_col = "peptides_counts_all",
  species = "Homo Sapiens"
)
}
\arguments{
\item{data}{The input data set from which categorization is performed on and hitlist is produced from;
i.e. the dataset after normalization}

\item{data_diff}{The output from imprints_caldiff; can be NULL and if so will compute it; can also be a path to the data.}

\item{ctrl}{The name of the control.}

\item{QP_name}{The name of the QP columns; default is 36C}

\item{FC_cutoff}{The fold change cutoff.}

\item{FDR}{The FDR used for the BH corrected combined p-value}

\item{curvature}{The curvature used for the curve on the volcano plot}

\item{folder_name}{The name of the folder in which you want to save the results.}

\item{peptide_count_col}{The name of the column that contain the unique peptide count.
If it is sumUniPeps, don't bother with this parameter.}

\item{species}{The species on which you did the experiment (not necessary if already present in your datas).
Default is 'Homo Sapiens'.}
}
\value{
A dataframe which contains the hits.
}
\description{
Function to categorize protein according to their Expression change
}
