% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_QPpep_rawread.R
\name{imprints_QPpep_rawread}
\alias{imprints_QPpep_rawread}
\title{imprints_QPpep_rawread}
\usage{
imprints_QPpep_rawread(data, treatment, contaminant = "Cont")
}
\arguments{
\item{data}{Your input data. It can the path to the output file from PD or a 
data.frame corresponding to that output.}

\item{treatment}{The treatment corresponding to each fo your TMT channel. It should be 
the same length as your number of TMT channel and should follow the format 'Rep_treatment'
like 'B1_Vehicle' for example.}

\item{contanimant}{Character to tell which suffix indicating a protein is a contaminant.}
}
\value{
Your data filtered and in the right format to use the IMPRINTS.CETSA package.
}
\description{
Function to read the quantitative proteomics data on the peptide level from Proteome Discoverer
}
